Add files via upload
This is code that I followed from freeCodeCamp on YouTube. This seems like a good way to get familiar with GitHub and hopefully increase my coding skills. The pictures were changed from the original. Pictures were taken by Charles Deluvio at Unsplash.com. I learned how to resize a picture on a MAC, which is new for me. There are extra pictures in this file as I was unsure of the number of pictures I needed at the time.